Reduce compilation time of epee/portable_storage.h
