+++ contrib/epee/include/storages/portable_storage_template_helper.h
@@ -32,6 +32,7 @@
 #include "parserse_base_utils.h"
 #include "portable_storage.h"
 #include "file_io_utils.h"
+#include "span.h"
 
 namespace epee
 {
