+++ contrib/epee/include/storages/portable_storage_to_bin.h
@@ -32,6 +32,7 @@
 #include "misc_language.h"
 #include "portable_storage_base.h"
 #include "portable_storage_bin_utils.h"
+#include "misc_log_ex.h"
 
 namespace epee
 {
